+++ snuba/datasets/storages/discover.py
@@ -29,17 +29,15 @@ from snuba.query.processors.physical.empty_tag_condition_processor import (
 from snuba.query.processors.physical.events_bool_contexts import (
     EventsBooleanContextsProcessor,
 )
+from snuba.query.processors.physical.hexint_column_processor import (
+    HexIntColumnProcessor,
+)
 from snuba.query.processors.physical.mapping_optimizer import MappingOptimizer
 from snuba.query.processors.physical.mapping_promoter import MappingColumnPromoter
 from snuba.query.processors.physical.null_column_caster import NullColumnCaster
 from snuba.query.processors.physical.prewhere import PrewhereProcessor
 from snuba.query.processors.physical.table_rate_limit import TableRateLimit
-from snuba.query.processors.physical.type_converters.hexint_column_processor import (
-    HexIntColumnProcessor,
-)
-from snuba.query.processors.physical.type_converters.uuid_column_processor import (
-    UUIDColumnProcessor,
-)
+from snuba.query.processors.physical.uuid_column_processor import UUIDColumnProcessor
 
 columns = ColumnSet(
     [
