+++ snuba/datasets/storages/errors_v2.py
@@ -25,6 +25,9 @@ from snuba.query.processors.physical.empty_tag_condition_processor import (
 from snuba.query.processors.physical.events_bool_contexts import (
     EventsBooleanContextsProcessor,
 )
+from snuba.query.processors.physical.hexint_column_processor import (
+    HexIntColumnProcessor,
+)
 from snuba.query.processors.physical.mapping_optimizer import MappingOptimizer
 from snuba.query.processors.physical.mapping_promoter import MappingColumnPromoter
 from snuba.query.processors.physical.prewhere import PrewhereProcessor
@@ -37,19 +40,14 @@ from snuba.query.processors.physical.tuple_unaliaser import TupleUnaliaser
 from snuba.query.processors.physical.type_condition_optimizer import (
     TypeConditionOptimizer,
 )
-from snuba.query.processors.physical.type_converters.hexint_column_processor import (
-    HexIntColumnProcessor,
-)
-from snuba.query.processors.physical.type_converters.uuid_array_column_processor import (
-    UUIDArrayColumnProcessor,
-)
-from snuba.query.processors.physical.type_converters.uuid_column_processor import (
-    UUIDColumnProcessor,
-)
 from snuba.query.processors.physical.uniq_in_select_and_having import (
     UniqInSelectAndHavingProcessor,
 )
 from snuba.query.processors.physical.user_column_processor import UserColumnProcessor
+from snuba.query.processors.physical.uuid_array_column_processor import (
+    UUIDArrayColumnProcessor,
+)
+from snuba.query.processors.physical.uuid_column_processor import UUIDColumnProcessor
 from snuba.replacers.replacer_processor import ReplacerState
 from snuba.subscriptions.utils import SchedulingWatermarkMode
 from snuba.utils.streams.topics import Topic
