+++ snuba/datasets/storages/transactions_common.py
@@ -24,21 +24,19 @@ from snuba.query.processors.physical.empty_tag_condition_processor import (
 from snuba.query.processors.physical.events_bool_contexts import (
     EventsBooleanContextsProcessor,
 )
+from snuba.query.processors.physical.hexint_column_processor import (
+    HexIntArrayColumnProcessor,
+    HexIntColumnProcessor,
+)
 from snuba.query.processors.physical.mapping_optimizer import MappingOptimizer
 from snuba.query.processors.physical.mapping_promoter import MappingColumnPromoter
 from snuba.query.processors.physical.prewhere import PrewhereProcessor
 from snuba.query.processors.physical.table_rate_limit import TableRateLimit
 from snuba.query.processors.physical.tuple_unaliaser import TupleUnaliaser
-from snuba.query.processors.physical.type_converters.hexint_column_processor import (
-    HexIntArrayColumnProcessor,
-    HexIntColumnProcessor,
-)
-from snuba.query.processors.physical.type_converters.uuid_column_processor import (
-    UUIDColumnProcessor,
-)
 from snuba.query.processors.physical.uniq_in_select_and_having import (
     UniqInSelectAndHavingProcessor,
 )
+from snuba.query.processors.physical.uuid_column_processor import UUIDColumnProcessor
 
 columns = ColumnSet(
     [
