+++ tests/query/processors/test_fixedstring_array_column_processor.py
@@ -6,7 +6,7 @@ from snuba.clickhouse.query import Query
 from snuba.query import SelectedExpression
 from snuba.query.data_source.simple import Table
 from snuba.query.expressions import Column, Expression, FunctionCall, Literal
-from snuba.query.processors.physical.type_converters.fixedstring_array_column_processor import (
+from snuba.query.processors.physical.fixedstring_array_column_processor import (
     FixedStringArrayColumnProcessor,
 )
 from snuba.query.query_settings import HTTPQuerySettings
